Broke out Sprint 2018.10 changes to mine.py model file into new class files
